requirements.txt
openai-agents
python-dotenv

# pip install -r requirements.txt